refactor: move oneliner display and add-flow helpers to executor_oneliners_ui.go
